changed the column type rate from integer to float

+++ db/migrate/20180809064120_create_taxation_obligations.rb
@@ -2,7 +2,7 @@ class CreateTaxationObligations < ActiveRecord::Migration[5.2]
   def change
     create_table :taxation_obligations do |t|
       t.string :type, null: false
-      t.integer :rate
+      t.float :rate
       t.references :taxation_form, foreign_key: true
       t.references :activity_type, foreign_key: true
